Remove conversion from size_t to int

+++ include/envoy/http/header_map.h
@@ -574,14 +574,14 @@ public:
    * @param key supplies the header key to remove.
    * @return the number of headers removed.
    */
-  virtual int remove(const LowerCaseString& key) PURE;
+  virtual size_t remove(const LowerCaseString& key) PURE;
 
   /**
    * Remove all instances of headers where the key begins with the supplied prefix.
    * @param prefix supplies the prefix to match header keys against.
    * @return the number of headers removed.
    */
-  virtual int removePrefix(const LowerCaseString& prefix) PURE;
+  virtual size_t removePrefix(const LowerCaseString& prefix) PURE;
 
   /**
    * @return the number of headers in the map.
